% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/structure.R
\name{applyStructure}
\alias{applyStructure}
\title{Apply structure of an R object}
\usage{
applyStructure(inst, data, scope, ...)
}
\arguments{
\item{inst}{A class instance.}
}
\value{
An error. See respective methods.
}
\description{
Designed for usage with implementations of the
\code{\link[idata]{IDataFormat}} interface and methods
\code{\link[idata]{IDataFormat$getStructure}}
}
\details{
TODO
}
\examples{

# data.frame --------------------------------------------------------------

inst <- data.frame(a = 1, b = 2, c = 3)
getStructure(inst)

# list --------------------------------------------------------------------

inst <- list(a = 1, b = 2, c = 3)
getStructure(inst)

# character ---------------------------------------------------------------

inst <- letters
getStructure(inst)
names(inst) <- letters
getStructure(inst)

}

