% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataCon.IntelligentForecaster.R
\docType{data}
\name{DataCon.IntelligentForecaster.Csv}
\alias{DataCon.IntelligentForecaster.Csv}
\title{Connector for Intelligent Forecaster CSV format}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x000000000ac6d0c0> 
 - attr(*, "name")= chr "DataCon.IntelligentForecaster.Csv_generator"
}}
\usage{
DataCon.IntelligentForecaster.Csv
}
\description{
This class wraps objects for connecting to data stored by
Intelligent Forecaster in a CSV format and provides respective methods.
}
\details{
TODO
}
\section{Fields}{

\describe{
\item{\code{con}}{\code{\link[base]{character}}
File path}

\item{\code{cached}}{\code{\link[base]{data.frame}}
 Cached data state as data frame}
}}
\section{Getters/setters}{


\itemize{
 \item{See superclass} {
   \code{\link[idata]{DataCon}}
 }
}
}

\section{Methods}{


\itemize{
 \item{See superclass} {
   \code{\link[idata]{DataCon}}
 }
}
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}
\keyword{datasets}

