% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataCon.IntelligentForecaster.R
\docType{data}
\name{DataCon.IntelligentForecaster}
\alias{DataCon.IntelligentForecaster}
\title{Connector for Intelligent Forecaster}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x0000000009c36378> 
 - attr(*, "name")= chr "DataCon.IntelligentForecaster_generator"
}}
\usage{
DataCon.IntelligentForecaster
}
\description{
This class wraps objects for connecting to Intelligent Forecaster and
provides respective methods.
}
\details{
TODO
}
\section{Fields}{

\describe{
\item{\code{con}}{\code{\link{character}}
 File path}
}}
\section{Getters/setters}{


\itemize{
 \item{See superclass} {
   \code{\link[idata]{DataCon}}
 }
}
}

\section{Public methods}{


\itemize{
 \item{See superclass} {
   \code{\link[idata]{DataCon}}
 }
}
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}
\keyword{datasets}

