% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataCon.Neo4j.R
\docType{data}
\name{DataCon.Neo4j}
\alias{DataCon.Neo4j}
\title{Connector for Neo4j}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x0000000009b39a48> 
 - attr(*, "name")= chr "DataCon.Neo4j_generator"
}}
\usage{
DataCon.Neo4j
}
\description{
This class wraps objects for connecting to Neo4j and provides
respective methods.
}
\details{
TODO
}
\section{Fields}{

\describe{
\item{\code{con}}{\code{\link{graph}}
Connection object}

\item{\code{cached}}{\code{\link[base]{data.frame}}
 Cached data state as data frame}
}}
\section{Public methods}{


\itemize{
 \item{See superclass} {
   \code{\link[idata]{IDataCon}}
 }
}
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}
\keyword{datasets}

