% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2_super_classes.R
\docType{data}
\name{DataCon}
\alias{DataCon}
\title{Superclass for inheritance for data connections}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x0000000009bb4038> 
 - attr(*, "name")= chr "DataCon_generator"
}}
\usage{
DataCon
}
\description{
TODO
}
\details{
TODO
}
\section{Fields}{

\describe{
\item{\code{con}}{\code{\link{ANY}}
Connection to a data source}

\item{\code{cached}}{\code{\link{ANY}}
 Cached data state}
}}
\section{Getters/setters}{


\itemize{
 \item{See interface} {
   \code{\link[idata]{IDataCon}}
 }
}
}

\section{Public methods}{


\itemize{
 \item{See interface} {
   \code{\link[idata]{IDataCon}}
 }
}
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}
\keyword{datasets}

