% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/2_super_classes.R
\docType{data}
\name{DataFormat}
\alias{DataFormat}
\title{Generic class for inheritance for data format information}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x0000000009c71e08> 
 - attr(*, "name")= chr "DataFormat_generator"
}}
\usage{
DataFormat
}
\description{
TODO
}
\details{
TODO
}
\section{Fields}{

\describe{
\item{\code{format}}{\code{\link{list}}
 Data format information}
}}
\section{Getters/setters}{


\itemize{
 \item{See superclass} {
   \code{\link[idata]{IDataFormat}}
 }
}
}

\section{Public methods}{


\itemize{
 \item{See interface} {
   \code{\link[idata]{IDataFormat}}
 }
}
}
\examples{

# Example formats ----------------------------------------------------------

format_1 <- list(
  function(x, pattern = "\\\\d{4}-\\\\d{2}-\\\\d{2}") {
    tmp <- lapply(x, function(ii) {
      if (any(grepl(pattern, ii))) {
        as.POSIXlt(ii)
      } else {
        ii
      }
    })
    as.data.frame(tmp, stringsAsFactors = FALSE)
  },
  function(x) {
    names(x) <- gsub("\\\\.", "_", names(x))
    x
  }
)
format_2 <- list(
  function(x, pattern = "\\\\d{4}-\\\\d{2}-\\\\d{2}") {
    tmp <- lapply(x, function(ii) {
      if (any(grepl(pattern, ii))) {
        as.character(ii)
      } else {
        ii
      }
    })
    as.data.frame(tmp, stringsAsFactors = FALSE)
  },
  function(x) {
    names(x) <- gsub("_", ".", names(x))
    x
  }
)

# Instantiate -------------------------------------------------------------

format_1_inst <- DataFormat$new(format = format_1)
format_2_inst <- DataFormat$new(format = format_2)

# Methods -----------------------------------------------------------------

format_1_inst$getFormat()
format_1_inst$setFormat(NULL)
format_1_inst$getFormat()
format_1_inst$setFormat(format_2)
format_1_inst$getFormat()

format_2_inst$getFormat()
format_2_inst$setFormat(NULL)
format_2_inst$getFormat()
format_2_inst$setFormat(format_1)
format_2_inst$getFormat()

}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}
\keyword{datasets}

