% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{Datasource.Neo4j}
\alias{Datasource.Neo4j}
\title{Connector for Neo4j}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x000000000b0c6db0> 
 - attr(*, "name")= chr "Datasource.Neo4j_generator"
}}
\usage{
Datasource.Neo4j
}
\description{
This class wraps objects for connecting to Neo4j and provides
respective methods.
}
\details{
TODO
}
\section{Fields}{

\describe{
\item{\code{con}}{\code{\link{graph}}
 Connection object}
}}
\section{Methods}{


\itemize{
 \item{See superclass} {
   \code{\link[idata]{IDatasource}}
 }
}
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}
\keyword{datasets}

