% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{Datasource}
\alias{Datasource}
\title{Generic class that specific classes for data sources inherit from}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x000000000b0eab38> 
 - attr(*, "name")= chr "Datasource_generator"
}}
\usage{
Datasource
}
\description{
TODO
}
\details{
TODO
}
\section{Fields}{

\describe{
\item{\code{con}}{\code{\link{ANY}}
 Connection to a data source}
}}
\section{Methods}{


\itemize{
 \item{initialize} {

   \itemize{
     \item{con }{\code{\link{ANY}}}
   }
 }
}
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}
\keyword{datasets}

