% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/structure.R
\name{handleSetStructure.default}
\alias{handleSetStructure.default}
\title{Handler for setting structure information}
\usage{
\method{handleSetStructure}{default}(inst)
}
\arguments{
\item{inst}{\code{\link[base]{ANY}}.}
}
\value{
\code{\link[base]{Structure}}. Return value of
 \code{\link[idata]{getStructure}}.
}
\description{
See generic: \code{\link[idata]{handleSetStructure}}
Method for: \code{\link[base]{ANY}}
}
\details{
TODO
}
\examples{

# handleSetStructure.default ----------------------------------------------

data <- data.frame(a = 1, b = 2, c = 3)
handleSetStructure(data)

# handleSetStructure.Structure --------------------------------------------

struc <- getStructure(data)
handleSetStructure(struc)
}

