% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/structure.R
\name{handleSetStructure}
\alias{handleSetStructure}
\title{Handler for setting structure information}
\usage{
handleSetStructure(inst, ...)
}
\arguments{
\item{inst}{A class instance.}
}
\value{
See respective methods.
}
\description{
Designed for usage with implementations of the
\code{\link[idata]{IDataFormat}} interface and method
\code{\link[idata]{IDataFormat$setStructure}}
}
\details{
TODO
}
\examples{

# handleSetStructure.default ----------------------------------------------

data <- data.frame(a = 1, b = 2, c = 3)
handleSetStructure(data)

# handleSetStructure.Structure --------------------------------------------

struc <- getStructure(data)
handleSetStructure(struc)
}

