% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/structure.R
\name{handleSetStructure.Structure}
\alias{handleSetStructure.Structure}
\title{Handler for setting structure information}
\usage{
\method{handleSetStructure}{Structure}(inst)
}
\arguments{
\item{inst}{\code{\link[base]{Structure}}.}
}
\value{
\code{\link[base]{Structure}}. Just returns input.
}
\description{
See generic: \code{\link[idata]{handleSetStructure}}
Method for: \code{\link[base]{Structure}}
}
\details{
TODO
}
\examples{

# handleSetStructure.default ----------------------------------------------

data <- data.frame(a = 1, b = 2, c = 3)
handleSetStructure(data)

# handleSetStructure.Structure --------------------------------------------

struc <- getStructure(data)
handleSetStructure(struc)
}

