% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idata-package.r
\docType{package}
\name{idata}
\alias{idata}
\alias{idata-package}
\title{Interfaces for data sources}
\description{
Interfaces for various data sources and related packages.
}
\details{
TODO
}
\section{Classes and constructors}{


\itemize{
   \item{\code{\link[idata]{TODO}}: }{

     TODO
   }
}
}

\section{Main functions/methods}{


 \itemize{
   \item{\code{\link[idata]{TODO}}: }{

     TODO
   }
}
}

\section{Disclaimer}{

\strong{This package is work-in-progress and is thus likely to change quite
substantially until it reaches a more mature state}
}

\section{Aknowledgements}{

This package was greatly inspired by the following resources:

\itemize{
   \item{\url{TODO}: } {
     TODO
   }
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}

