% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_classes.R
\docType{data}
\name{IDataCon}
\alias{IDataCon}
\title{Class that functions as an interface for data connections}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x000000000acf5ca0> 
 - attr(*, "name")= chr "IDataCon_generator"
}}
\usage{
IDataCon
}
\description{
Defines the interface for data connections.
}
\details{
The terms \emph{interace} is used in a looser context than in more
rigid OOP contexts such as \emph{C#} or the like
}
\section{Getters/setters}{


\itemize{
 \item{getCached} {
 }
 \item{getCachedActive} {
 }
 \item{setCached} {
 }
 \item{setCachedActive} {
 }
}
}

\section{Public methods}{


\itemize{
 \item{initialize} {
 }
 \item{toExternalFormat} {
 }
 \item{toRFormat} {
 }
 \item{pull} {
 }
 \item{push} {
 }
}
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}
\keyword{datasets}

