% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_interface_classes.R
\docType{data}
\name{IDataFormat}
\alias{IDataFormat}
\title{Class that functions as an interface for data format information}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x0000000009c801b0> 
 - attr(*, "name")= chr "IDataFormat_generator"
}}
\usage{
IDataFormat
}
\description{
Defines the interface for data format information.
}
\section{Details on interface term}{

The term \emph{interface} is used in a looser context than in more
rigid OOP contexts such as \emph{C#} or the like
}

\section{Getters/setters}{


\itemize{
 \item{getFormat} {
   TODO
 }
 \item{setFormat} {
   TODO
 }
 \item{getStructure} {
   TODO
 }
 \item{setStructure} {
   TODO
 }
}
}

\section{Public methods}{


\itemize{
 TODO
}
}

\section{Known interface implementations}{


\itemize{
 \item{\code{\link[idata]{DataFormat}}}
}
}
\examples{

# Example formats ----------------------------------------------------------

format_1 <- list(
  function(x, pattern = "\\\\d{4}-\\\\d{2}-\\\\d{2}") {
    tmp <- lapply(x, function(ii) {
      if (any(grepl(pattern, ii))) {
        as.POSIXlt(ii)
      } else {
        ii
      }
    })
    as.data.frame(tmp, stringsAsFactors = FALSE)
  },
  function(x) {
    names(x) <- gsub("\\\\.", "_", names(x))
    x
  }
)
format_2 <- list(
  function(x, pattern = "\\\\d{4}-\\\\d{2}-\\\\d{2}") {
    tmp <- lapply(x, function(ii) {
      if (any(grepl(pattern, ii))) {
        as.character(ii)
      } else {
        ii
      }
    })
    as.data.frame(tmp, stringsAsFactors = FALSE)
  },
  function(x) {
    names(x) <- gsub("_", ".", names(x))
    x
  }
)

# Instantiate -------------------------------------------------------------

format_1_inst <- DataFormat$new(format = format_1)
format_2_inst <- DataFormat$new(format = format_2)

# Methods -----------------------------------------------------------------

format_1_inst$getFormat()
format_1_inst$setFormat(NULL)
format_1_inst$getFormat()
format_1_inst$setFormat(format_2)
format_1_inst$getFormat()

format_2_inst$getFormat()
format_2_inst$setFormat(NULL)
format_2_inst$getFormat()
format_2_inst$setFormat(format_1)
format_2_inst$getFormat()

}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}
\keyword{datasets}

