% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{IDatasource}
\alias{IDatasource}
\title{General interface class that other interfaces inherit from/implement}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x000000000b114260> 
 - attr(*, "name")= chr "IDatasource_generator"
}}
\usage{
IDatasource
}
\description{
This is the class that defines the overall interfaces that is inherited
from/implemented by other interfaces.
}
\details{
The terms \emph{interace} is used in a more loose context than in more
rigid OOP contexts such as \emph{C#} or the like
}
\section{Fields}{

\describe{
\item{\code{con}}{\code{\link{ANY}}
 Connection to a data source}
}}
\section{Methods}{


\itemize{
 \item{initialize} {

   \itemize{
     \item{con }{\code{\link{ANY}}}
   }
 }
}
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}
\keyword{datasets}

