% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{mapDatasourceToList.Datasource.Neo4j}
\alias{mapDatasourceToList.Datasource.Neo4j}
\title{Map Neo4j database to list}
\usage{
\method{mapDatasourceToList}{Datasource.Neo4j}(con)
}
\arguments{
\item{con}{\code{\link[RNeo4j]{Datasource.Neo4j}}.
Neo4j graph wrapped by connector class.}
}
\value{
\code{\link[base]{list}}.
}
\description{
Maps Neo4j database as provided by package \code{RNeo4j}
(\url{https://cran.rstudio.com/web/packages/RNeo4j/index.html}) and
wrapped by connector class \code{\link[datar]{Datasource.Neo4j}}
to a \code{\link{list}}.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}

