% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{mapDatasourceToList.graph}
\alias{mapDatasourceToList.graph}
\title{Map Neo4j database to list}
\usage{
\method{mapDatasourceToList}{graph}(con)
}
\arguments{
\item{con}{\code{\link[RNeo4j]{graph}}.
Neo4j graph.}
}
\value{
\code{\link[base]{list}}.
}
\description{
Maps Neo4j database as provided by package \code{RNeo4j}
(\url{https://cran.rstudio.com/web/packages/RNeo4j/index.html})
to \code{\link{list}}.
}
\details{
TODO
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}

