% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/code.R
\name{mapDirectory}
\alias{mapDirectory}
\title{Map directory structure}
\usage{
mapDirectory(dir = ".", include_root = TRUE, include_files = TRUE,
  relative = FALSE)
}
\arguments{
\item{dir}{\code{\link[base]{character}}.
Directory path.}

\item{include_root}{\code{\link[base]{logical}}.
\code{TRUE}: include root directory;
\code{FALSE}: do not include root directory. See examples.}

\item{include_files}{\code{\link[base]{logical}}.
\code{TRUE}: include actual files;
\code{FALSE}: do not include files. See examples.}

\item{include_files}{\code{\link[base]{logical}}.
\code{TRUE}: include actual files;
\code{FALSE}: do not include files. See examples.}
}
\value{
\code{\link[base]{list}}.
}
\description{
Maps directory structure to a \code{\link{list}}.
}
\details{
Whether or not only directories or actual files should be included can
be controlled by \code{include_files}.
}
\examples{

# Example 1 ---------------------------------------------------------------

mapDirectory("R")
mapDirectory("R", include_root = FALSE)
mapDirectory("R", include_files = FALSE)


# Example 2 ---------------------------------------------------------------

tmp <- tempdir()
paths <- file.path(tmp, c("a/b/c", "b"))
sapply(paths, dir.create, recursive = TRUE, showWarnings = FALSE)
file.create(file.path(tmp, c("a/a_1.txt", "a/a_2.txt")))
file.create(file.path(tmp, c("a/b/c/c_1.txt", "a/b/c/c_2.txt")))

mapDirectory(dir = file.path(tmp, "a"))
mapDirectory(dir = file.path(tmp, "a"), include_root = FALSE)
mapDirectory(dir = file.path(tmp, "a"), include_files = FALSE)

mapDirectory(dir = tmp)


# Example 3 ---------------------------------------------------------------

mapDirectory(dir = tmp, relative = TRUE)
mapDirectory(dir = tmp, include_root = FALSE, relative = TRUE)
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}

