% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validate.R
\name{methodNotImplemented.R6}
\alias{methodNotImplemented.R6}
\title{Stop on interface call}
\usage{
\method{methodNotImplemented}{R6}(inst, cl = sys.call(-2))
}
\arguments{
\item{inst}{An instance of class \code{\link[R6]{R6}}.}
}
\value{
An error informing the user that the method he called
 has not yet been implemented.
}
\description{
See generic: \code{\link[idata]{methodNotImplemented}}.

Method for instances of class \code{\link[R6]{R6}}
}
\details{
TODO
}
\examples{
## TODO
}

