% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validate.R
\name{methodNotImplemented}
\alias{methodNotImplemented}
\title{Method not implemented yet}
\usage{
methodNotImplemented(inst, ...)
}
\arguments{
\item{inst}{A class instance. Designed for use in class methods.}
}
\value{
An error. See respective methods.
}
\description{
Used inside methods that have not been (fully) implemented yet.
}
\details{
TODO
}
\examples{
## TODO
}

