% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pull.R
\name{pullFromCon.DataCon.IntelligentForecaster.Csv}
\alias{pullFromCon.DataCon.IntelligentForecaster.Csv}
\title{Pull from a connection}
\usage{
\method{pullFromCon}{DataCon.IntelligentForecaster.Csv}(con)
}
\arguments{
\item{con}{\code{\link[idata]{DataCon.IntelligentForecaster.Csv}}.}
}
\value{
\code{\link[base]{data.frame}}.
}
\description{
Pulls data from a connection object of class
\code{\link[idata]{DataCon.IntelligentForecaster.Csv}}
}
\details{
TODO
}
\examples{
## TODO
}

