% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pull.R
\name{pullFromCon}
\alias{pullFromCon}
\title{Pull from a connection}
\usage{
pullFromCon(con)
}
\arguments{
\item{con}{A connection object}
}
\value{
See respective methods.
}
\description{
Pulls data from a connection.
}
\details{
TODO
}
\examples{
## TODO
}

