% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validate.R
\name{stopOnInterfaceCall.R6}
\alias{stopOnInterfaceCall.R6}
\title{Stop on interface call}
\usage{
\method{stopOnInterfaceCall}{R6}(inst, cl = sys.call(-2))
}
\arguments{
\item{inst}{An instance of class \code{\link[R6]{R6}}.}
}
\value{
An error informing the user that he called an interface method.
}
\description{
See generic: \code{\link[idata]{stopOnInterfaceCall}}.

Method for instances of class \code{\link[R6]{R6}}
}
\details{
TODO
}
\examples{
## TODO
}

