% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/validate.R
\name{stopOnInterfaceCall}
\alias{stopOnInterfaceCall}
\title{Stop on interface call}
\usage{
stopOnInterfaceCall(inst, ...)
}
\arguments{
\item{inst}{A class instance. Designed for instances of interface classes.}
}
\value{
An error. See respective methods.
}
\description{
Used inside methods of interface classes to signal that the interface
instead of an interface \emph{implementation} has been called.
}
\details{
TODO
}
\examples{
## TODO
}

