% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_generics.R
\name{toExternalFormat}
\alias{toExternalFormat}
\title{Format to external format}
\usage{
toExternalFormat(con)
}
\arguments{
\item{con}{A connection object}
}
\value{
See respective methods.
}
\description{
Formats data to external format.
}
\details{
TODO
}
\examples{
## TODO
}

