% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataCon.IntelligentForecaster.R
\name{toRFormat.DataCon.IntelligentForecaster.Csv}
\alias{toRFormat.DataCon.IntelligentForecaster.Csv}
\title{Format to R format}
\usage{
\method{toRFormat}{DataCon.IntelligentForecaster.Csv}(con, date_col = "date",
  date_format = "\%m/\%d/\%Y \%H:\%M:\%S", extended = FALSE,
  with_ids = FALSE, ...)
}
\arguments{
\item{con}{\code{\link[idata]{DataCon.IntelligentForecaster.Csv}}.}
}
\value{
Formated \code{\link[base]{data.frame}}.
}
\description{
Formats data to R format for objects of class
\code{\link[idata]{DataCon.IntelligentForecaster.Csv}}
}
\details{
TODO
}
\examples{
## TODO
}

