% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format.R
\name{toRFormat}
\alias{toRFormat}
\title{Format to R format}
\usage{
toRFormat(con)
}
\arguments{
\item{con}{A connection object}
}
\value{
See respective methods.
}
\description{
Formats data to R format.
}
\details{
TODO
}
\examples{
## TODO
}

