% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/classes.R
\docType{data}
\name{Datasource.Neo4j}
\alias{Datasource.Neo4j}
\title{Connector for Neo4j}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x000000000b116c28> 
 - attr(*, "name")= chr "Datasource.Neo4j_generator"
}}
\usage{
Datasource.Neo4j
}
\description{
Main class to handle communication with Neo4j
}
\details{
TODO
}
\section{Fields}{

\describe{
\item{\code{cache}}{\code{\link{character}}
 File path to cached R object.
 Takes precedence over other meta information if specified}
}}
\section{Methods}{


\itemize{
 \item{initialize} {

   \itemize{
     \item{path }{\code{\link{character}}}
   }
 }
}
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}
\keyword{datasets}

