% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/1_interface_classes.R
\docType{data}
\name{IDataCon}
\alias{IDataCon}
\title{Class that functions as an interface for data connections}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x0000000009c32c70> 
 - attr(*, "name")= chr "IDataCon_generator"
}}
\usage{
IDataCon
}
\description{
Defines the interface for data connections.
}
\section{Details on interface term}{

The term \emph{interface} is used in a looser context than in more
rigid OOP contexts such as \emph{C#} or the like
}

\section{Getters/setters}{


\itemize{
 \item{getCached} {
   TODO
 }
 \item{getCachedActive} {
   TODO
 }
 \item{setCached} {
   TODO
 }
 \item{setCachedActive} {
   TODO
 }
}
}

\section{Public methods}{


\itemize{
 \item{applyExternalFormat} {
   TODO
 }
 \item{applyRFormat} {
   TODO
 }
 \item{pull} {
   TODO
 }
 \item{push} {
   TODO
 }
}
}

\section{Known interface implementations}{


\itemize{
 \item{\code{\link[idata]{DataCon}}}
 \item{\code{\link[idata]{DataCon.IntelligentForecaster}}}
 \item{\code{\link[idata]{DataCon.IntelligentForecaster.Csv}}}
}
}
\examples{
## TODO
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/datar}
}
\keyword{datasets}

