% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/format.R
\name{toExternalFormat.DataCon.Neo4j}
\alias{toExternalFormat.DataCon.Neo4j}
\title{Format to external format}
\usage{
\method{toExternalFormat}{DataCon.Neo4j}(con)
}
\arguments{
\item{con}{\code{\link[idata]{DataCon.Neo4j}}.}
}
\value{
Formated \code{\link[base]{data.frame}}.
}
\description{
Formats data to external format for objects of class
\code{\link[idata]{DataCon.Neo4j}}
}
\details{
TODO
}
\examples{
## TODO
}

